Seems to compile now

+++ Tuber/src/main/java/com/parse/tuber/Profile.java
@@ -185,7 +185,7 @@ public class Profile extends Activity implements View.OnClickListener {
 
                     if (user != null) {
                         tvName.setText(user.get("name").toString());
-                        tvPrice.setText(String.format("$%d/hr", user.get("fee")));
+                        tvPrice.setText(String.format("$%.2f/hr", (double) user.get("fee")));
                         if (user.getBoolean("tutor") == false) {
                             tvPrice.setVisibility(View.GONE);
                             tvPriceLabel.setVisibility(View.GONE);
